requirements.txt
Flask-WTF
WTForms
Flask-SQLAlchemy
flask-migrate
flask-shell-ipython
flask-login
flask-bootstrap
flask-moment
python-dotenv
boto3
requests
authlib
six
auth0-python
python3-linkedin
flask-admin
flask-principal